WIP: Add Lines and Station
Lines will be used to check directions and stations on a given line.
Station can be used to check the next stops on a given station without
having to add it to the "stations to check" list.

The UI can still be improved. This is just a very quick and dirty
implementation.